+++ modules/safebrowsing.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 import time
 from dotenv import dotenv_values
-from list_utils import chunks
-from logger_utils import init_logger
-from ray_utils import execute_with_ray
+from modules.list_utils import chunks
+from modules.logger_utils import init_logger
+from modules.ray_utils import execute_with_ray
 import requests
 from requests.models import Response
 import itertools
@@ -11,7 +11,7 @@ import logging
 from tqdm import tqdm
 import base64
 
-from requests_utils import get_with_retries, post_with_retries
+from modules.requests_utils import get_with_retries, post_with_retries
 
 GOOGLE_API_KEY = dotenv_values(".env")["GOOGLE_API_KEY"]
 YANDEX_API_KEY = dotenv_values(".env")["YANDEX_API_KEY"]
